requirements.txt
Django==1.5.4
argparse==1.2.1
distribute==0.6.24
django-bootstrap-toolkit==2.15.0
pyredmine==0.2.4
python-dateutil==2.1
six==1.4.1
wsgiref==0.1.2
